+++ lib/appium_lib_core/device.rb
@@ -40,7 +40,7 @@ module Appium
             :stop_recording_screen, :stop_and_save_recording_screen,
             :shake, :device_time,
             :touch_actions, :multi_touch,
-            :execute_driver
+            :execute_driver, :execute_cdp
           ].each(&method(:delegate_from_appium_driver))
         end