+++ test/unit/android/device/w3c/commands_test.rb
@@ -480,6 +480,28 @@ class AppiumLibCoreTest
             assert_requested(:post, "#{SESSION}/element", times: 1)
             assert_requested(:post, "#{SESSION}/element/element_id_parent/element", times: 1)
           end
+
+          def test_chromium_send_command
+            stub_request(:post, "#{SESSION}/goog/cdp/execute")
+              .with(body: { cmd: 'Page.captureScreenshot', params: { quality: 1, format: 'jpeg' } }.to_json)
+              .to_return(headers: HEADER, status: 200, body: { value: { data: '/9j/4AAQSkZJRgABAQAAAQABAAD' } }.to_json)
+
+            r = @driver.execute_cdp 'Page.captureScreenshot', { quality: 1, format: 'jpeg' }
+
+            assert_requested(:post, "#{SESSION}/goog/cdp/execute", times: 1)
+            assert_equal '/9j/4AAQSkZJRgABAQAAAQABAAD', r['data']
+          end
+
+          def test_chromium_send_command_no_param
+            stub_request(:post, "#{SESSION}/goog/cdp/execute")
+              .with(body: { cmd: 'Page.getResourceTree', params: {} }.to_json)
+              .to_return(headers: HEADER, status: 200, body: { value: { frameTree: { childFrames: [] } } }.to_json)
+
+            r = @driver.execute_cdp 'Page.getResourceTree'
+
+            assert_requested(:post, "#{SESSION}/goog/cdp/execute", times: 1)
+            assert_equal({ 'childFrames' => [] }, r['frameTree'])
+          end
         end # class CommandsTest
       end # module W3C
     end # module Device